fix param request

+++ DataAccess/Repositories/ProductRepository.cs
@@ -15,7 +15,7 @@ namespace DataAccess.Repositories
 	{
 		public void AddProduct(Product product) =>  ProductDAO.Instance.AddProduct(product);
 
-		public List<SellerProductResponseDTO> GetAllProduct(int userId) => ProductDAO.Instance.GetAllProduct(userId);
+		public List<SellerProductResponseDTO> GetAllProduct(long userId) => ProductDAO.Instance.GetAllProduct(userId);
 
         public List<AllProductResponseDTO> GetAllProduct() => ProductDAO.Instance.GetAllProduct();
